Limit polling namespaces in API container
Removes the "compute" polling namespace from the ceilometer-polling
agent inside the API container.

By default both the "compute" and "central" polling namespaces are
enabled [1], and the compute namespace requires libvirt to be
installed or else ceilometer-polling will fail to start [2]. Since
we are not hosting instances inside the API container, we can
disable the namespace.

[1]: https://github.com/openstack/ceilometer/blob/4108d6d2cbe2ae509ed532b0be5c0f7d405fd0d6/ceilometer/cmd/polling.py#L65
[2]: http://logs.openstack.org/22/391522/2/check/gate-openstack-ansible-openstack-ansible-aio-ubuntu-xenial/6837eba/logs/openstack/aio1_ceilometer_api_container-2218d6eb/ceilometer/ceilometer-polling.log.txt.gz

Change-Id: I2046eb03bf26c62ac822d9a81847417bfcf24fb1